* PluginUpdates:
  Update JetPack
  Update Reveal IDs
  Add OAuth.
  Remove generics from themes.
  Remove Page Link To
  Delete unused plugins: Adminer NextGen Gallery
  Plugin updates: Admin Management Xtended Admin Post Navigation Disable Comments Explanatory Dictionary Media Tags Network Latest Posts Page Links To Print Friendly and PDF Which Template WP-Piwik

+++ wp-content/plugins/explanatory-dictionary/uninstall.php
@@ -1,12 +1,10 @@
 <?php
 /**
- * Fired when the plugin is uninstalled.
- *
- * @package   Dictionary_List
- * @author    Robert-John van Doesburg <info@rjvandoesburg.com>
+ * @package   Explanatory_Dictionary
+ * @author    EXED internet (RJvD, BHdH) <service@exed.nl>
  * @license   GPL-2.0+
- * @link      http://example.com
- * @copyright 2013 Robert-John van Doesburg
+ * @link      http://www.mixcom.nl/online
+ * @copyright 2014  EXED internet  (email : service@exed.nl)
  */
 
 // If uninstall, not called from WordPress, then exit
@@ -14,13 +12,14 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-// TODO: Define uninstall functionality here
-
-global $wpdb;
-$table_name = $wpdb->prefix . "dictionary_list";
-$sql = "
-	DROP TABLE `{$table_name}`
-";
-	
-require_once( ABSPATH . 'wp-admin/includes/upgrade.php' );
-dbDelta( $sql );
+require_once( plugin_dir_path( __FILE__ ) . 'public/classes/default.php' );
+if( Explanatory_Dictionary::VERSION < '4.1.5') {
+	global $wpdb;
+	$table_name = $wpdb->prefix . "dictionary_list";
+	$sql = "
+		DROP TABLE `{$table_name}`
+	";
+		
+	require_once( ABSPATH . 'wp-admin/includes/upgrade.php' );
+	dbDelta( $sql );
+}
